[refactor]pages폴더를 만들어 page별로 page와 component 구분
