+++ fe/src/pages/Account/SignUp/index.jsx
@@ -9,7 +9,7 @@ import {
   SignUpContainer,
   SignupHeader,
 } from './style';
-import API from '../../../../../lib/api';
+import API from 'lib/api';
 
 const SignUp = () => {
   const isLogined = useContext(IsLoginedState);
