+++ fe/src/pages/Picker/style.js
