+++ fe/src/pages/index.jsx
@@ -1,0 +1,10 @@
+export { default as Calendar } from 'pages/Calendar';
+export { default as Login } from 'pages/Account/Login';
+export { default as FreeBoard } from 'pages/FreeBoard';
+export { default as SignUp } from 'pages/Account/SignUp';
+export { default as Profile } from 'pages/Account/Profile';
+export { default as Edit } from 'pages/Account/Edit';
+export { default as Search } from 'pages/Search';
+export { default as Picker } from 'pages/Picker';
+export { default as Callback } from 'pages/Account/Login/Callback';
+export { default as Write } from 'pages/Write';
